// packages/vi-lib/src/vctp-parsing/vctp-attribute-type.ts
export enum VCTPAttributeType {
  Unknown = 0,
  NumberFlag = 0x1,
  StringFlag = 0x31,
  ArrayDimensions = 0x41,
  ArrayFixedSize = 0x42,
  ClusterFormat = 0x51,
  ClusterFormatStr = 0x52,
  RefType = 0x71,
  RefTypeName = 0x72,
  RefControlFlags = 0x73,
  RefEventRegistFlags = 0x74,
  RefDataValFlags = 0x75,
  TerminalPattern = 0xF1,
  TerminalFlags = 0xF2,
  TypDefFlag1 = 0xF5,
  TypDefControlNameCount = 0xF6,
  TypDefControlName1 = 0xF7,
  TypDefControlName2 = 0xF8,
  TypDefControlName3 = 0xF9,
}